Improvement on premake files - Project Working in wsl2-linux under clang-19 & gcc-10 (All Configurations) [clang-19] General Requirements are to define LLVMInstallDir in ~./bashrc & override the default symbols (ar, ld, etc...) Downgraded to opengl 4.5 (Wsl2 mesa drivers Don't support opengl 4.6) Rewrote util/file.hpp because fopen_s wasn't fully supported cross-platform Rewrote util/util_api.hpp, supporting Cygwin, MinGW, Linux, Clang, Windows TODO Automated testing of different platforms TODO Automated testing of each project for compilation & intended behaviour

+++ projects/util/include/util/file.hpp
@@ -13,9 +13,9 @@ namespace util {
 	After which the function will proceed with file loading
 */
 UTIL_API bool loadFile(
-	const char* 		path, /* filepath 														   */ 
-	unsigned long long* size, /* the size of the out array that was pre-allocated.                 */
-	char*				out   /* Where to place the file contents (Buffer of Min-Size 'size')      */
+	const char*   path, /* filepath 													*/ 
+	unsigned int* size, /* the size of the out array that was pre-allocated.            */
+	char*		  out   /* Where to place the file contents (Buffer of Min-Size 'size') */
 );
